requirements.txt
streamlit
pandas
requests
python-dotenv
yfinance
transformers
torch
matplotlib
requests-cache
lxml
sqlalchemy
psycopg2-binary
python-dotenv
tqdm